pyproject.toml
[tool.poetry]
name = "agentx"
version = "0.1.0"
description = "An AI controller that relies on three LLM chains to perform tasks"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
langchain = "^0.0.144"
openai = "^0.27.4"
pydantic = "^1.10.7"
faiss-cpu = "^1.7.3"
tiktoken = "^0.3.3"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
python-minifier = "^2.8.1"
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[tool.isort]
profile = "black"
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
